% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_functions.r
\name{autoUpS}
\alias{autoUpS}
\title{Automatically update Settings}
\usage{
autoUpS()
}
\description{
Use this function within your code to automatically update the 
settings from the user´s settings file
}
