% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_functions.r
\name{copyUnisetTemplates}
\alias{copyUnisetTemplates}
\title{Copy Uniset templates}
\usage{
copyUnisetTemplates(target = "~")
}
\arguments{
\item{target}{Where the templates should be copied to. Provide either a valid
path to a folder, or leave at the default '~' to copy the files to your home 
directory.}
}
\description{
Copies the required templates (a file with global variables
and the template for the settings.r file) either to the "home" directory,
or to any desired folder.
}
\details{
If the usual '~' for the users home directory can not be expanded, 
the files are instead copied to 'user/home'. Existing files will be overwritten.
The 'settings.r' file has to be moved into the 'inst' folder (create one if 
not already done) of your package, while the 'unisetGlobals.r' file goes into 
the 'R' folder of your package.
}
\examples{
\dontrun{
copyUnisetTemplates() # to copy them to your home directory
copyUnisetTemplates("foo/moo/bar") # copy the templates to folder 'bar'
}
}
