% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_functions.R
\name{uniset_autoUpS}
\alias{uniset_autoUpS}
\title{Automatically update Settings}
\usage{
uniset_autoUpS(unisetEnv)
}
\arguments{
\item{unisetEnv}{Character length one. Hand over the global variable defined 
in the target package holding the name of the uniset-environment for the 
specific target package ('uniset_env_name' or 'uev', see examples at 
\code{\link{uniset}}.)}
}
\description{
Use this function within your code to automatically update the 
settings from the users settings file
}
\details{
If 'autoUpdateSettings' in xxx_settings.r is left at 'TRUE', the 
settings will be checked resp. updated automatically every time a function in 
the target package is calling \code{\link{uniset_autoUpS}}.
}
