% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_functions.R
\name{uniset_copyFilesToPackage}
\alias{uniset_copyFilesToPackage}
\title{Copy Uniset Files into Target Package}
\usage{
uniset_copyFilesToPackage(
  pathToPackage,
  taPaEnv = "def",
  taPaObj = "stn",
  taPaSH = "def",
  tmpl = "_TEMPLATE"
)
}
\arguments{
\item{pathToPackage}{Character length one. The path to the root of the
target package.}

\item{taPaEnv}{Character length one. The name of the environment where the
settings for the target package (in a key=value format) will be stored. It is
recommended to use a rather short name starting with a '.' (dot). See details.
If left at the default 'def', the first two characters of the package name,
prepended with a '.' (dot), and appended with an 'e' (for environment) will
be used.}

\item{taPaObj}{Character length one. The name of the object (residing in the
environment with the name provided in argument 'taPaEnv') holding the list
with all the individual key-value pairs that can be defined to be used in the
target package. It is recommended to use a rather short name. See details.
Defaults to 'stn'.}

\item{taPaSH}{Character length one. The name of the variable to be defined in
the '.Renviron' file, leading to the place where the settings.r file for the
target package will be stored. If left at the default 'def', \code{taPaName_SH}
will be used, with 'taPaName' being the value provided at the argument
'taPaName'.}

\item{tmpl}{Character length one. the Character string that will be appended
to the fresh settings file that is possibly copied (by the target package)
to the users settings home directory. Can be left at the default '_TEMPLATE'.}
}
\value{
Writes the three required files directly into a valid R-package folder
structure. Returns (invisible) NULL.
}
\description{
Function to generate the three files required in the target
package (i.e. the package that should be enabled to use the package 'uniset').
The generated files will be copied directly into their required destination
folders in the target package. The name of the target package will be extracted
from the description file.
}
\details{
The importance of providing rather short characters at the arguments
'taPaEnv' and 'taPaObj' lies in the fact that those two names will be used to
get any parameter stored in the settings file of the target package. For example,
if you provide '.doe' at the argument 'taPaEnv' and 'stn' at the argument
'taPaObj', then all the values stored in the settings file can be retrieved by
calling \code{.doe$stn$XXX}, with 'XXX' being any of the keys defined in the
settings file.
}
\examples{
\dontrun{
# for an imaginary package called 'dogPack':
aa <- "~/desktop/dogPack"
uniset_copyFilesToPackage(aa) # uses the defaults for all arguments
uniset_copyFilesToPackage(aa, ".dpe", "sn")
	# see details.
}
}
\seealso{
\code{\link{uniset_getFiles}}
}
