% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_functions.r
\name{uniset_test2}
\alias{uniset_test2}
\title{Test2}
\usage{
uniset_test2(uev = get("uniset_env_name"))
}
\arguments{
\item{uev}{Character length one. The name of the environment holding the uniset 
definitions for a single package.}
}
\description{
Test2
}
