% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_functions.R
\name{uniset_autoUpS}
\alias{uniset_autoUpS}
\title{Automatically update Settings}
\usage{
uniset_autoUpS()
}
\description{
Use this function within your code to automatically update the 
settings from the users settings file
}
\details{
If 'autoUpdateSettings' in xxx_settings.r is left at 'TRUE', the 
settings will be checked resp. updated automatically every time a function in 
the target package is calling \code{\link{uniset_autoUpS}}.
}
