% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_functions.R
\name{uniset_autoUpS}
\alias{uniset_autoUpS}
\title{Automatically update Settings}
\usage{
uniset_autoUpS()
}
\description{
Use this function within your code to automatically update the 
settings from the users settings file
}
