% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_functions.r
\name{uniset_test}
\alias{uniset_test}
\title{Test}
\usage{
uniset_test(unisetEnv)
}
\arguments{
\item{Character}{length one. The name of the environment holding the uniset 
definitions for a single package.}
}
\description{
Test
}
