% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_functions.R
\name{uniset_test}
\alias{uniset_test}
\title{Simple Test}
\usage{
uniset_test(unisetEnv)
}
\arguments{
\item{unisetEnv}{Character length one. The name of the environment holding 
the uniset definitions for a single package.}
}
\value{
Is printing the parameters defined by the target package, and is 
returning those parameters in an (invisible) list.
}
\description{
Test if the input package name etc. was correct / successful. 
This function is meant to be called from inside the  target package.
}
\section{Important}{
 This function is intended to be called from within the 
target package.
}

\section{Note}{
 Please refer to \code{\link{uniset}} for a link to examples 
and a real-world demo.
}

\examples{
{
\dontrun{
# to be called from within the target package
uniset_test(get("uev"))
}
}
}
