% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen_functions.R
\name{uniset_test}
\alias{uniset_test}
\title{Simple Test}
\usage{
uniset_test(uniset_handover)
}
\arguments{
\item{uniset_handover}{List length two, containing two elements:
\enumerate{
\item \code{pkgname}: The name of the target package.
\item \code{funcname}: The name of the function handing over the required values.
}}
}
\value{
Is printing the parameters defined by the target package, and is
returning those parameters in an (invisible) list.
}
\description{
Test if the input package name etc. was correct / successful.
This function is meant to be called from inside the  target package.
}
\section{Important}{
 This function is intended to be called from within the
target package.
}

\section{Note}{
 Please refer to \code{\link{uniset}} for a link to examples
and a real-world demo.
}

\examples{
{
\dontrun{
# to be called from within the target package
uniset_test(uniset_handover_pkgname, uniset_handover_funcname)
}
}
}
